refactor, removing special case around appKey - also generalize what we where trying to do there - pass in config. so now it's `SecretStack(defaults).use(...)...(config)`
